% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQEval.data.R
\docType{data}
\name{AQEval.data}
\alias{AQEval.data}
\alias{aq.data}
\title{AQEval Example data}
\format{
(26280x6) 'tbl_df' objects
\describe{
  \item{date}{Time-series of POSIX class date and time records.}
  \item{no2}{Time-series of
  nitrogen dioxide measurements from local site.}
  \item{bg.no2}{Time-series of
  nitrogen dioxide measurements from nearby
  background site.}
  \item{ws}{Time-series of
  local wind speed measurements.}
  \item{wd}{Time-series of
  local wind direction measurements.}
  \item{air_temp}{Time-series
  of local air temperature measurements.}
}
}
\source{
Air quality and meteorological data packaged
for use with AQEval Examples.

aq.my1 air quality data from Marylebone Road AURN
monitoring station.

aq.ea2 and aq.kc1 air quality data from Earling and
Knightsbridge KCL network monitoring stations.

Data access from associated archives using
\code{\link{openair}}
functions \code{\link{importAURN}} and
\code{\link{importKCL}} and \code{\link{worldmet}}
function \code{\link{importNOAA}}.
}
\usage{
aq.data
}
\description{
Data packaged with AQEval for
use with example code.
}
\references{
Carslaw, D. C. and K. Ropkins, (2012) openair --- an
R package for air quality data analysis.
Environmental Modelling & Software. Volume 27-28,
52-61.

David Carslaw (2021). worldmet: Import Surface
Meteorological Data from NOAA Integrated Surface
Database (ISD). R package version 0.9.5.
\url{https://CRAN.R-project.org/package=worldmet}
}
\keyword{datasets}
