% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqe.generics.R
\name{aqeval.generics}
\alias{aqeval.generics}
\alias{print.aqe}
\alias{aqe.generics}
\alias{plot.aqe}
\title{aqeval.generics}
\usage{
\method{print}{aqe}(x, ...)

\method{plot}{aqe}(x, ...)
}
\arguments{
\item{x}{the \code{aqe}
object to be printed, plotted, etc.}

\item{...}{additional arguments, typically passed on to next method or
ignored.}
}
\description{
Generic functions for use with \code{aqe} object class for
\code{AQEval} outputs.
}
\keyword{internal}
