% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQEval-package.R
\docType{package}
\name{AQEval}
\alias{AQEval}
\title{Air Quality Evaluation}
\description{
R AQEval beta-release package of R code for Air
Quality time-series discrete change analysis.
}
\section{TO DO}{


Need to standardise wording here and intro paper...

ref intro paper?

acknowledgement to funders?
}

\section{main package functions}{

\code{\link{findBreakPoints}},
\code{\link{quantBreakPoints}},
\code{\link{quantBreakSegments}}.
}

