% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqRAG.R
\name{aqRAG}
\alias{aqRAG}
\alias{RAG03}
\alias{fitDSWModel}
\alias{aqiRAG}
\alias{makeIntTimeSeries}
\title{Air Quality Red Amber Green (RAG)}
\usage{
RAG03(
  data,
  pollutant = "no2",
  to = NULL,
  from = NULL,
  mod = NULL,
  pad.data = NULL,
  x.args = NULL,
  ...,
  only.last = TRUE,
  out = "all"
)

aqiRAG(
  data,
  pollutant = "no2",
  intervention = NULL,
  rag = NULL,
  mrag = NULL,
  threshold = 40,
  to = NULL,
  mod = NULL,
  ...,
  output = c("report", "scores")
)

fitDSWModel(
  data,
  pollutant = "no2",
  to = NULL,
  from = NULL,
  mod = NULL,
  pad.data = NULL,
  x.args = NULL,
  ...,
  only.last = TRUE,
  out = "all"
)

makeIntTimeSeries(data, from = NULL, to = NULL, int.descr = NULL, ...)
}
\arguments{
\item{data}{(data.frame or tibble) data source.}

\item{pollutant}{(character) name of species RAG model
is to be built for.}

\item{to}{(various) date to extrapolate model to, by
default end of supplied data date range.}

\item{from}{(various) date to extrapolate model from,
by default start of supplied data date range.}

\item{mod}{(character) the left-hand-side of the RAG
model if not using standard formula.}

\item{pad.data}{(various) data to pad extrapolated
regions with.}

\item{x.args}{(character vector) Names of x terms to
use in model.}

\item{...}{Extra arguments.}

\item{only.last}{(logical) Option to just report final
value when calculation rolling annual mean.}

\item{out}{(vector) Function output type, by default
all model outputs.}

\item{intervention}{(data.frame or similar) time-series
of intervention contributions.}

\item{int.descr}{(vector) predicted intervention
impacts. (NEEDS DOC)}
}
\description{
working function used in the development
of the RAG methods.
}
\note{
this is a work function, not intended to be final
version and not fully document - please handle with
care.
}
