% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalStepImpact.R
\name{evalStepImpact}
\alias{evalStepImpact}
\title{evalStepImpact}
\usage{
evalStepImpact(
  mod,
  pollutant = NULL,
  start = NULL,
  end = NULL,
  breaks = NULL,
  segments = NULL,
  ...
)
}
\arguments{
\item{mod}{Data source, typically a quantBreakSegments output.}

\item{pollutant}{the pollutant to build the step impact estimate
for, if not the one the break-point/segment model was built for.
This needs a WARNING...}

\item{start}{Start of time period impact of step-change to be
evaluated for (YYYY-MM-DD), by default from start of data
range.}

\item{end}{End of time period impact of step-change to be
evaluated for (YYYY-MM-DD), by default from end of data
range.}

\item{breaks}{The break-points to evaluate impact of.}

\item{segments}{The segments to evaluate impact of.}

\item{...}{Other arguments, currently ignored}
}
\description{
Estimates time period impact of a step-change.
}
