% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreakPoints.R
\name{findBreakPoints}
\alias{findBreakPoints}
\title{findBreakPoints}
\usage{
findBreakPoints(data, pollutant, h = 0.15, ...)
}
\arguments{
\item{data}{dataset containing time-series to be tested,
identified using pollutant, and a reference time-stamp
called 'date'}

\item{pollutant}{name of time-series}

\item{h}{data window to use as reference when looking for
breaks from previous trends, expressed as proportion of
time-series (0-1)}

\item{...}{other parameters}
}
\description{
find break points in a time-series using
strucchange methods
}
\note{
This uses methods from strucchange package (see references) and
modified as suggested by the main author of strucchange to handle
missing cases. URL:
https://stackoverflow.com/questions/43243548/strucchange-not-reporting-breakdates
}
\references{
Strucchange:

Achim Zeileis, Friedrich Leisch, Kurt Hornik and Christian Kleiber
(2002). strucchange: An R Package for Testing for Structural Change
in Linear Regression Models. Journal of Statistical Software, 7(2),
1-38. URL http://www.jstatsoft.org/v07/i02/

Achim Zeileis, Christian Kleiber, Walter Kraemer and Kurt Hornik
(2003). Testing and Dating of Structural Changes in Practice.
Computational Statistics & Data Analysis, 44, 109-123.
}
