% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNearSiteModel.R
\name{fitNearSiteModel}
\alias{fitNearSiteModel}
\title{Fit a near-site model}
\usage{
fitNearSiteModel(data, pollutant = "no2", y, x = "rest", elements = NULL, ...)
}
\arguments{
\item{data}{data.frame (or similar) containing data-series
to be modelled; this is expected to contain 'date', 'site'
and pollutant of interest data-series}

\item{pollutant}{The pollutant of interest, by default 'NO2'}

\item{y}{The site to be modelled}

\item{x}{The sites to use when building the model, the
default 'rest' uses all supplied sites except 'y'}

\item{elements}{The number of inputs to use in the
site models, can be any number up to length of x or
combination thereof; by default, length(x):1}
}
\description{
Build an air quality model for one location
using air quality data from nearby sites
}
