% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateContribution.R
\name{isolateContribution}
\alias{isolateContribution}
\title{isolateContribution}
\usage{
isolateContribution(
  data,
  pollutant,
  background = NULL,
  deseason = TRUE,
  deweather = TRUE,
  method = 2,
  add.term = NULL,
  formula = NULL,
  output = "mean",
  ...
)
}
\arguments{
\item{data}{data source, typically data.frame, containing time-series}

\item{pollutant}{the time-series}

\item{background}{(optimal) if supplied, the background time-series to
subtract}

\item{deseason}{logical, if TRUE (default) the data is deseasonalised as
part of contribution isolation}

\item{deweather}{logical, if TRUE (default) the data is deweathered as
part of contribution isolation}

\item{method}{numeric, contribution isolation method (default 2)}

\item{add.term}{extra terms to add to the contribution isolation
model; ignore for now (in development)}

\item{formula}{model formula to use; ignore for now (in development)}

\item{output}{output options; ignore for now (in development)}

\item{...}{other arguments; ignore for now (in development)}
}
\description{
time-series contribution isolation based on
background subtraction, deseasonalisation and deweathering
}
