% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantBreakPoints.R
\name{quantBreakPoints}
\alias{quantBreakPoints}
\title{quantBreakPoints}
\usage{
quantBreakPoints(
  data,
  pollutant,
  breaks,
  ylab = NULL,
  xlab = NULL,
  pt.col = c("lightgrey", "darkgrey"),
  line.col = "red",
  break.col = "blue",
  show = c("plot", "report"),
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar,
containing data-series to model and a paired timestamp
data-series, named date.}

\item{pollutant}{The data-series to model.}

\item{breaks}{The break points and confidence intervals,
typically a findBreakPoints function output.}

\item{ylab}{Y-label term, by default pollutant.}

\item{xlab}{X-label term, by default date.}

\item{pt.col}{Point fill and line colours for plot,
defaults lightgrey and darkgrey.}

\item{line.col}{Line colour for plot, default red.}

\item{break.col}{Break point colour for plot, default
blue.}

\item{show}{What to show before returning the break-point
quantification mode, by default plot and report.}

\item{...}{other parameters}
}
\description{
Quantifies Break Points, assuming a
model based on general trend plus any supplied break
points.
}
\note{
quantBreakPoints is the default version of this function;
quantBreaks01 and quantBreaks02, now removed, were early versions used in
testing and optimisation work.
}
\examples{
#using openair timeAverage
temp <- openair::timeAverage(aq.my1, "4 day")
quantBreakPoints(temp, "no2")
}
