% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantBreakSegments.R
\name{quantBreakSegments}
\alias{quantBreakSegments}
\title{quantBreakSegments}
\usage{
quantBreakSegments(
  data,
  pollutant,
  breaks,
  ylab = NULL,
  xlab = NULL,
  pt.col = c("lightgrey", "darkgrey"),
  line.col = "red",
  break.col = "blue",
  seg.method = 2,
  show = c("plot", "report"),
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar,
containing data-series to model and a paired timestamp
data-series, named date.}

\item{pollutant}{The data-series to model.}

\item{breaks}{The break points and confidence intervals,
typically a findBreakPoints function output.}

\item{ylab}{Y-label term, by default pollutant.}

\item{xlab}{X-label term, by default date.}

\item{pt.col}{Point fill and line colours for plot,
defaults lightgrey and darkgrey.}

\item{line.col}{Line colour for plot, default red.}

\item{break.col}{Break transition colour for plot, default
blue.}

\item{seg.method}{Method to use to fit segments: 1 back-stop;
2 soft segmented (default).}

\item{show}{What to show before returning the brak-point
quantification mode, by default plot and report.}

\item{...}{other parameters}
}
\description{
Assumes model based on general trend plus
transitions about declared break points in supplied data.
}
\note{
quantBreakSegments is the default version of this function;
quantSegments01 and quantSegments02 were early versions used in testing
and optimisation work.
}
