% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBreakPoint.R
\name{testBreakPoints}
\alias{testBreakPoints}
\title{testBreakPoints}
\usage{
testBreakPoints(data, pollutant, breaks, ...)
}
\arguments{
\item{data}{dataset containing time-series to be tested,
and a reference time-stamp called 'date'}

\item{pollutant}{name of time-series to be tested}

\item{breaks}{The break points and confidence intervals,
typically a findBreakPoints function output}

\item{...}{other parameters}
}
\description{
test break points identified using
findBreakPoints
}
