% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQEval.data.R
\docType{data}
\name{AQEval.data}
\alias{AQEval.data}
\alias{aq.data}
\title{AQEval Example data}
\format{
(26280x6) 'tbl_df' objects
\describe{
  \item{date}{Time-series of POSIX class date and time records.}
  \item{no2}{Time-series of
  nitrogen dioxide measurements from local site.}
  \item{bg.no2}{Time-series of
  nitrogen dioxide measurements from nearby
  background site.}
  \item{ws}{Time-series of
  local wind speed measurements.}
  \item{wd}{Time-series of
  local wind direction measurements.}
  \item{air_temp}{Time-series
  of local air temperature measurements.}
}
}
\source{
Air quality and meteorological data packaged
for use with AQEval Examples.

Time-series sources:
\itemize{
 \item no2 Nitrogen dioxide downloaded from King's
 College London Archive using \code{importKCL}
 function in \code{openair}.
 \item bg.no2  Nitrogen dioxide downloaded from
 Autonmatic Urban and Rural Network Archive using
 \code{importAURN} function in \code{openair}.
 \item ws, wd, air_temp Wind speed, wind direction
 and air temperature downloaded from NOAA's
 Integrated Surface Database using \code{importNOAA}
 function in \code{worldmet}.
}
}
\usage{
aq.data
}
\description{
Data packaged with AQEval for
use with example code.
}
\references{
Carslaw, D. C. and K. Ropkins, (2012) openair --- an
R package for air quality data analysis.
Environmental Modelling & Software. Volume 27-28,
52-61.

David Carslaw (2021). worldmet: Import Surface
Meteorological Data from NOAA Integrated Surface
Database (ISD). R package version 0.9.5.
\url{https://CRAN.R-project.org/package=worldmet}
}
\seealso{
\code{\link{openair}}: functions \code{\link{importAURN}} and
\code{\link{importKCL}}

\code{\link{worldmet}}: function \code{\link{importNOAA}}.
}
\keyword{datasets}
