% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQEval-package.R
\docType{package}
\name{AQEval}
\alias{AQEval}
\title{Air Quality Evaluation}
\description{
Pre-release package for R code for Air Quality Evaluation.
}
\section{main package functions}{

findBreakPoints, quantBreakPoints, quantBreakSegments
}

