% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.breaks.R
\name{quantify.breaks}
\alias{quantify.breaks}
\alias{quantBreakPoints}
\alias{quantBreakSegments}
\title{quantify break-point/segments}
\usage{
quantBreakPoints(
  data,
  pollutant,
  breaks,
  ylab = NULL,
  xlab = NULL,
  pt.col = c("lightgrey", "darkgrey"),
  line.col = "red",
  break.col = "blue",
  show = c("plot", "report"),
  ...
)

quantBreakSegments(
  data,
  pollutant,
  breaks,
  ylab = NULL,
  xlab = NULL,
  pt.col = c("lightgrey", "darkgrey"),
  line.col = "red",
  break.col = "blue",
  seg.method = 2,
  show = c("plot", "report"),
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar,
containing data-series to model and a paired time-stamp
data-series, named date.}

\item{pollutant}{The name of the data-series to
break-point or break-segment model.}

\item{breaks}{(Optional) The break-points to
and confidence intervals to use when build either
break-point or break-segment models. If not supplied
these are build using \code{\link{findBreakPoints}}
and supplied arguments.}

\item{ylab}{Y-label term, by default pollutant.}

\item{xlab}{X-label term, by default date.}

\item{pt.col}{Point fill and line colours for plot,
defaults lightgrey and darkgrey.}

\item{line.col}{Line colour for plot, default red.}

\item{break.col}{Break-point/segment colour for plot, default
blue.}

\item{show}{What to show before returning the break-point
quantification mode, by default plot and report.}

\item{...}{other parameters}

\item{seg.method}{(\code{quantBreakSegments} only) the
break-segment fitting method to use.}
}
\value{
Both functions use the \code{show} argument
to control which elements of the functions outputs
are shown but also invisible return a \code{list}
of all outputs which can caught using, e.g.:

\code{brk.mod <- quantBreakPoints(data, pollutant)}
}
\description{
Quantify either break-points or
break-segment methods for pollutant time-series
}
\details{
\code{quantBreakPoints} and
\code{quantBreakSegments} both use
\code{strucchange} methods to identify potential
break-points in time-series, and then quantify
these as conventional break-points or break-segments,
respectively.
}
\examples{
#using openair timeAverage to covert 1-hour data to 1-day averages
temp <- openair::timeAverage(aq.data, "1 day")

#break-points
quantBreakPoints(temp, "no2", h=0.3)

#break-segments
quantBreakSegments(temp, "no2", h=0.3)
}
\references{
Ropkins et al (In Prep).
}
\seealso{
\code{\link{timeAverage}} in \code{openair}.
}
\author{
Karl Ropkins
}
