% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralFrequency.R
\name{spectralFrequency}
\alias{spectralFrequency}
\title{Spectral Analysis}
\usage{
spectralFrequency(data, pollutant, ...)
}
\arguments{
\item{data}{data.frame holding data to be analysed,
expected to contain a timestamp data-series called
date and a measurement time-series to be analysed
identified using the pollutant argument}

\item{pollutant}{the name of the time-series, typically
pollutant measurements, to be analysed}

\item{...}{extra arguments}
}
\description{
Time-series spectral frequency analysis.
}
